Add Gradle build directory support
In this commit, support for the Gradle build directory is added. This
means that, instead of assuming the Maven ('target') build directory, it
is checked whether this exists. If it does not, instead, the Gradle
('build') build directory is assumed, and set if it exists. If neither
exists, then a FileNotFoundException is thrown.